requirements.txt
asgiref==3.7.2
boto3==1.33.9
botocore==1.33.9
Django==4.1.12
django-admin-soft-dashboard==1.0.12
django-storages==1.14.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.5.0
